Install python3 after starting the jail
Instead of installing python3 via the `chroot` option of pkgng before
the jail is actually started, we now install it after the jail started
via the `jail` option of pkgng.

This actually works while the other option did only because of some
manual commands that were run before.

Sadly, it does not bootstrap pkg inside of the jail and we still need to
run `pkg -y` once in the new jail (see package role).